requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
Pillow==10.2.0
PyInstaller==6.3.0
